Added create , read and updated dto

+++ WalletApp.Domain/Entities/Wallet.cs
@@ -5,7 +5,7 @@ namespace WalletApp.Domain.Entities;
 public class Wallet : BaseEntity
 {
     public decimal CurrentBalance { get; private set; }
-    public float TotalPoints { get; private set; }
+    public decimal TotalPoints { get; private set; }
     public Guid UserId { get; set; }
     public WalletUser User { get; set; }
